+++ setup/plummer.h
@@ -2,6 +2,6 @@
 #define PLUMMER_H_
 
 double g(double);
-void plummer_dist(double *, double *, double *, const int, const double, const double, const double);
+void plummer_dist(double *, double *, double *, const int, const int, const double, const double);
 
 #endif // PLUMER_H_
